Wait for subscription to begin testing
It can take some time for an asynchronous DDS implementation to
initialize and connect with existing subscribers. Before beginning the
round trip test, we should make sure that the ROS publisher has
successfully connected to the subscriber.

This should resolve the test failure when using Fast-DSS, which is the
new default RMW implementation for ROS 2 Humble. Without waiting, the
first message being published from ROS seems to be lost, meaning the
first message recieved from Drake is a '0', which fails the test.

Signed-off-by: Scott K Logan <logans@cottsay.net>