Add suppression for cpplint in Humble
The name of the rule which tracks this behavior has changed in the
version of cpplint used in Humble. Unfortunately, there doesn't seem to
be a single suppression that works with both Galactic and Humble
cpplint.

This filter should unblock builds in Humble, while the existing
suppressions mentioned in the comment should continue to work for
Galactic.

Signed-off-by: Scott K Logan <logans@cottsay.net>