Update Bazelized build to use the latest ROS 2 Rolling underlay archive (#71)
Switch bazelized builds to use drake-ros-underlay tarball

The rolling-on-focal tarball is no longer built, and we should switch to
the newer purpose-built drake-ros-underlay tarball. The tarball is
rebuilt often - at least weekly, so manually curating the sha256 of the
archive isn't sustainable. Rather, securely download a checksum file to
use when validating the archive.

Drop rosidl_typesupport_c__visibility_control.h
This header file was dropped in a rosidl_typesupport_c 2.0.0.

Update ROS 2 prerequisite rosdep key list
The list of rosdeps required for the ROS 2 packages we're targeting has
changed a bit over the past several weeks.

Drop prefix chaining workaround
This workaround is no longer needed when using the drake-ros-underlay
archive instead of the rolling-on-focal archive.

Signed-off-by: Scott K Logan <logans@cottsay.net>