Stop including entire ament index for each package
The ament index design specs state that marker files in the resource
index should share the name of the package installing them. We can
leverage this to select which marker files belong to each package.

The existing approach worked for isolated workspace layouts, but broke
down when inspecting a merged workspace.