[cmake] Turn off -Werror=extra -Werror=pedantic (#360)
In no universe should Werror be enabled in end-user builds (rather
than just developers local builds and CI checks), and certainly not
for warning categories that Drake's own CI doesn't even enable. We
don't want *this* repository to be a warning-catcher for Drake code.

The easy fix for now is to just turn off the overly broad warning
categories. This does not fix the bug that Werror=all is enabled for
builds triggered by downstream users.