Fix over-escaping by CMake
When the workspace path has ~ characters (and possibly others), the JSON
output by CMake will be a quoted string, e.g.

```
"lib": "\"/some/ws~~path/libfoo.so\""
```

The extract path will be `"/some/ws~~path/libfoo.so"` instead of
`/some/ws~~path/libfoo.so` and the former will not be treated as path
down the line.