Miscellaneous scrapping fixes (#378)
* Remove elements that have a condition attribute on ROS1

* Fix over-escaping by CMake

When the workspace path has ~ characters (and possibly others), the JSON
output by CMake will be a quoted string, e.g.

```
"lib": "\"/some/ws~~path/libfoo.so\""
```

The extract path will be `"/some/ws~~path/libfoo.so"` instead of
`/some/ws~~path/libfoo.so` and the former will not be treated as path
down the line.

* Include *.inc files in C++ headers

This is used in the domain_bridge package.

* Wrap to 80 characters

* Address review comments

* Handle the case where top_level.txt has more than one entry

* Correctly handle multiple top level entries

* Fix typo in top_levels expression